+++ lib/chef-vault/item.rb
@@ -104,12 +104,27 @@ class ChefVault::Item < Chef::DataBagItem
     end
   end
 
-  def rotate_keys!
+  def rotate_keys!(clean_unknown_clients=false)
     @secret = generate_secret
 
     unless clients.empty?
-      clients.each do |client|
-        clients("name:#{client}")
+      if clean_unknown_clients
+        clients_to_remove=[]
+        clients.each do |client|
+          begin
+            clients("name:#{client}")
+          rescue ChefVault::Exceptions::ClientNotFound
+            clients_to_remove.push(client)
+          end
+        end
+        clients_to_remove.each do |client|
+          puts "Removing unknown client '#{client}'"
+          clients("name:#{client}", :delete)
+        end
+      else
+        clients.each do |client|
+          clients("name:#{client}")
+        end
       end
     end