+++ Changelog.md
@@ -1,6 +1,10 @@
 ## Planned (Unreleased)
 
 ## Released
+## v2.4.0 / 2014-12-03
+* add simplecov test coverage configuration (Doug Ireton)
+* add --clean-unknown-clients switch to knife remove/rotate (Thomas Gschwind and Reto Hermann)
+
 ## v2.3.0 / 2014-10-22
 * add --clean switch to knife update (thanks to Matt Brimstone)
 * added aruba CLI testing framework (just for --clean option for now)
@@ -81,4 +85,3 @@
 * Add encrypt cert
 
 ## v0.1.1 / 2013-03-14
-